Library

Gadfly.plot — Method

plot(chn::MCMCChains.Chains,
    elements::Gadfly.ElementOrFunctionOrLayers...; mapping...) -> Plot

Plot a chains object by transforming it to a DataFrame. Settings are passed to Gadfly via elements and mapping.

Gadfly.render — Method

Gadfly.Geom.render(geom::DensityCI, theme, aes)

We need to do all the rendering manually because Gadfly.Geom.density is a line element and therefore cannot do a ribbon at the same time. The way to get that normally would be via Gadfly.Stat.density with Geom.polygon. Unfortunately, the polygon doesn't work with subplot_grid.

TuringPlots.apply_filter! — Method

apply_filter!(df, mapping)

If the user has defined a filter, then apply this to df.

TuringPlots.n_chains — Method

n_chains(chn::Chains, parameter)

Number of chains for chn[parameter].

TuringPlots.plot_parameters — Method

plot_parameters(chn::Chains)

Plot sample value and density for each parameter of chains. The plot is built by creating two Gadfly.subplot_grids and using Gadfly.hstack.